Move stuff to `window/rules.rs`
